requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
distlib==0.3.8
Django==5.1.6
django-taggit==6.1.0
filelock==3.15.4
idna==3.7
pillow==11.1.0
platformdirs==4.2.2
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.2
virtualenv==20.26.3
